pyproject.toml
[project]
name = "budgettracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.3",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "mysqlclient>=2.2.8",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
]
